fixing up display on multiple devices
