+++ src/components/Footer/style.css
@@ -1,5 +1,4 @@
 .footer {
-  margin-top: 10%;
   background-color: var(--offblack);
   color: var(--offwhite);
 } .footericons {
