+++ src/pages/Home/style.css
@@ -3,7 +3,56 @@
     --smokeygrey: #c3c0b7;
     --offblack: #060606;
     --offwhite: #e3e4e6;
+}
+@media screen and (min-width: 1921px) and (min-height: 1081px) { /* Desktop */
+  .content {
+    width: 100vw;
+    height: 100vh;
+    margin-top: 82px;
   }
-.Text {
-    margin-top: 108px;
+}
+@media screen and (max-width: 1920px) and (max-height: 1080px) { /* Desktop */
+  .content {
+    width: 100vw;
+    height: 100vh;
+    margin-top: 82px;
+  }
+}
+@media  screen and (max-width: 1024px) and (max-height: 1366px) { /* Tablet */
+  .content {
+    width: 100vw;
+    height: 200vw;
+    margin-top: 0px;
+  }
+}
+@media screen and (max-width: 430px) and (max-height: 932px) { /* Mobile */
+  .content {
+    width: 100vw;
+    height: 300vh;
+    margin-top: 0px;
+  }
+}
+.content {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  overflow: hidden;
+}
+.box1 {
+  position: relative;
+  background-color: red;
+  width: 100%;
+  height: 33.3%;
+}
+.box2 {
+  position: relative;
+  background-color: blue;
+  width: 100%;
+  height: 33.4%;
+}
+.box3 {
+  position: relative;
+  background-color: green;
+  width: 100%;
+  height: 33.3%;
 }
